make container exit with rc 1 if main nginx exits uncleanly
The previous solution, an infinite loop depending on existing pid file,
causes an infinite and fast restart loop in case of a repeated crash
(due to a mis-configuration or resource problem), which cannot be a good
solution. Instead, one should rely on the container orchestrator to take
care of a restart.